+++ src/vec2.hh
@@ -2,19 +2,19 @@
 #include <cmath>
 
 struct Vec2 {
-    float x;
-    float y;
-    Vec2(float _x, float _y): x(_x), y(_y) {};
+    double x;
+    double y;
+    Vec2(double _x, double _y): x(_x), y(_y) {};
     Vec2(): x(0), y(0) {};
-    float dot(Vec2 const & v) const {
+    double dot(Vec2 const & v) const {
         return x * v.x + y * v.y;
     }
-    float len() const {
+    double len() const {
         return std::sqrt(this->dot(*this));
     }
-    Vec2 rotate(float angle) {
-        float s = std::sin(angle);
-        float c = std::cos(angle);
+    Vec2 rotate(double angle) {
+        double s = std::sin(angle);
+        double c = std::cos(angle);
         return {x * c - y * s, x * s + y * c};
     }
 };
@@ -46,26 +46,26 @@ inline Vec2 & operator/=(Vec2 & lhs, Vec2 const & rhs) {
 
 template <typename T>
 inline Vec2 & operator+=(Vec2 & lhs, T rhs) {
-    lhs.x += (float) rhs;
-    lhs.y += (float) rhs;
+    lhs.x += (double) rhs;
+    lhs.y += (double) rhs;
     return lhs;
 }
 template <typename T>
 inline Vec2 & operator-=(Vec2 & lhs, T rhs) {
-    lhs.x -= (float) rhs;
-    lhs.y -= (float) rhs;
+    lhs.x -= (double) rhs;
+    lhs.y -= (double) rhs;
     return lhs;
 }
 template <typename T>
 inline Vec2 & operator*=(Vec2 & lhs, T rhs) {
-    lhs.x *= (float) rhs;
-    lhs.y *= (float) rhs;
+    lhs.x *= (double) rhs;
+    lhs.y *= (double) rhs;
     return lhs;
 }
 template <typename T>
 inline Vec2 & operator/=(Vec2 & lhs, T rhs) {
-    lhs.x /= (float) rhs;
-    lhs.y /= (float) rhs;
+    lhs.x /= (double) rhs;
+    lhs.y /= (double) rhs;
     return lhs;
 }
 
@@ -78,4 +78,4 @@ inline Vec2 operator*(Vec2 lhs, T rhs) {return lhs *= rhs;}
 template <typename T>
 inline Vec2 operator/(Vec2 lhs, T rhs) {return lhs /= rhs;}
 
-inline Vec2 operator*(float lhs, Vec2 rhs) {return rhs *= lhs;}
+inline Vec2 operator*(double lhs, Vec2 rhs) {return rhs *= lhs;}
